requirements.txt
requests>=2.28.0
Pillow>=9.0.0
pydub>=0.25.1
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0
static-ffmpeg>=2.5
